# Lab6Deploy/Building-a-Node.js-TypeScript-MySQL-Boilerplate-API-main/swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Sign-up and Verification API
  version: 1.0.0
  description: Node.js + MySQL - API with email sign-up, verification, authentication and forgot password
servers:
  - url: http://localhost:4000
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /accounts/authenticate:
    post:
      summary: Authenticate account credentials and return a JWT token and a cookie with a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: admin@example.com
                password:
                  type: string
                  example: asdf1234
      responses:
        '200':
          description: OK
  /accounts/refresh-token:
    post:
      summary: Use a refresh token to generate a new JWT token and a new refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
  /accounts/revoke-token:
    post:
      summary: Revoke a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
  /accounts/register:
    post:
      summary: Register a new user account and send a verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, firstName, lastName, email, password, confirmPassword, acceptTerms]
              properties:
                title:
                  type: string
                  example: Mr
                firstName:
                  type: string
                  example: Gabriel
                lastName:
                  type: string
                  example: Abellana
                email:
                  type: string
                  example: gab@example.com
                password:
                  type: string
                  example: asdf1234
                confirmPassword:
                  type: string
                  example: asdf1234
                acceptTerms:
                  type: boolean
                  example: true
      responses:
        '200':
          description: OK
  /accounts/verify-email:
    post:
      summary: Verify a new account with a verification token received by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
  /accounts/forgot-password:
    post:
      summary: Submit email address to reset the password on an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  example: gab@example.com
      responses:
        '200':
          description: OK
  /accounts/validate-reset-token:
    post:
      summary: Validate the reset token received by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
  /accounts/reset-password:
    post:
      summary: Reset the password on an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password, confirmPassword]
              properties:
                token:
                  type: string
                password:
                  type: string
                  example: asdf1234
                confirmPassword:
                  type: string
                  example: asdf1234
      responses:
        '200':
          description: OK
  /accounts:
    get:
      summary: Get a list of all accounts (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /accounts/{id}:
    get:
      summary: Get a single account by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      summary: Update an account
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
                confirmPassword:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: OK
    delete:
      summary: Delete an account
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK